feat: add support to check ETags
This check is added to avoid where one thread/process fetches an Entity from Cosmos and modifies that version in-memory and another thread/process updates the Entity before it is saved (updated) by the first. It is the built-in [optimistic concurrency control of Cosmos](https://docs.microsoft.com/en-us/azure/cosmos-db/sql/database-transactions-optimistic-concurrency).