devops: add a delay when failing to process dead letter message
We need to cater for when CosmosDb is completely down. Then a message could end up in dead letter due to the CosmosDb outage but the auto processing of dead letter messages (via ServiceBusTrigger) will also fail due to it wanting to dump it in CosmosDb too. This commit adds a delay when the command fails (command would fail if it cannot write to CosmosDb).